dags/requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.15.0
boto3==1.34.0
pandas==2.2.1